440: Split voter group checks into separate files
