+++ apps/backend/src/controllers/.bodyChecks/voterGroupChecks/common.check.ts
@@ -1,20 +1,13 @@
 import type { BallotPaper as DBBallotPaper, User as DBUser } from '@repo/db/types';
-import {
-  insertableVoterGroupObject,
-  updateableVoterGroupObject,
-  zodErrorToResponse400,
-  type InsertableVoterGroup,
-  type UpdateableVoterGroup,
-} from '@repo/votura-validators';
 import type { Selectable } from 'kysely';
-import { HttpStatusCode } from '../../httpStatusCode.js';
+import { HttpStatusCode } from '../../../httpStatusCode.js';
 import {
   checkBallotPapersBelongToUser,
   checkBallotPapersElectionNotFrozen,
   checkBallotPapersExist,
   checkBallotPapersFromDifferentElections,
-} from '../../services/ballotPapers.service.js';
-import type { BodyCheckValidationError } from './bodyCheckValidationError.js';
+} from '../../../services/ballotPapers.service.js';
+import type { BodyCheckValidationError } from '../bodyCheckValidationError.js';
 
 export enum VoterGroupValidationErrorMessage {
   ballotPaperNotFound = 'One or more ballot papers to be added to voter group not found.',
@@ -27,7 +20,7 @@ export interface VoterGroupValidationError extends BodyCheckValidationError {
   message: VoterGroupValidationErrorMessage | string;
 }
 
-const defaultVoterGroupChecks = async (
+export const defaultVoterGroupChecks = async (
   userId: Selectable<DBUser>['id'],
   ballotPaperIds: Selectable<DBBallotPaper>['id'][],
 ): Promise<VoterGroupValidationError | null> => {
@@ -69,44 +62,4 @@ const defaultVoterGroupChecks = async (
   }
 
   return null;
-};
-
-export const validateInsertableVoterGroup = async (
-  body: unknown,
-  userId: string,
-): Promise<InsertableVoterGroup | VoterGroupValidationError> => {
-  // Validate the request body against the insertableVoterGroupObject schema
-  const { data, error, success } = await insertableVoterGroupObject.safeParseAsync(body);
-  if (!success) {
-    return {
-      status: HttpStatusCode.badRequest,
-      message: zodErrorToResponse400(error).message,
-    };
-  }
-
-  const validationError = await defaultVoterGroupChecks(userId, data.ballotPapers);
-  if (validationError !== null) {
-    return validationError;
-  }
-  return data;
-};
-
-export const validateUpdateableVoterGroup = async (
-  body: unknown,
-  userId: string,
-): Promise<UpdateableVoterGroup | VoterGroupValidationError> => {
-  // Validate the request body against the updateableVoterGroupObject schema
-  const { data, error, success } = await updateableVoterGroupObject.safeParseAsync(body);
-  if (!success) {
-    return {
-      status: HttpStatusCode.badRequest,
-      message: zodErrorToResponse400(error).message,
-    };
-  }
-
-  const validationError = await defaultVoterGroupChecks(userId, data.ballotPapers);
-  if (validationError !== null) {
-    return validationError;
-  }
-  return data;
 };
