+++ apps/backend/test/voterGroups/createVoterGroup.test.ts
@@ -9,7 +9,7 @@ import request from 'supertest';
 import { beforeAll, describe, expect, it } from 'vitest';
 import { app } from '../../src/app.js';
 import { generateUserTokens } from '../../src/auth/utils.js';
-import { VoterGroupValidationErrorMessage } from '../../src/controllers/.bodyChecks/voterGroupChecks.js';
+import { VoterGroupValidationErrorMessage } from '../../src/controllers/.bodyChecks/voterGroupChecks/common.check.js';
 import { HttpStatusCode } from '../../src/httpStatusCode.js';
 import { createUser, findUserBy } from '../../src/services/users.service.js';
 import {
